Update tui printing, and fix bug with multiple games (see desc.)
Updated styling of tui-printing. Also moved game controller initialization to its own method, which is now called when user 'starts' a game, rather than once on app launch. This was due to games not resetting before, causing all games after the first one to immediately end due to presence of a winner.